+++ R/dendrogram.R
@@ -19,9 +19,16 @@
 
 #' Coersions to Dendrogram
 #'
-#' Coerce [hclust], [hdbscan] or [reachability] objects to [dendrogram].
+#' Provides a new generic function to coerce objects to dendrograms with
+#' [stats::as.dendrogram()] as the default. Additional methods for
+#' [hclust], [hdbscan] and [reachability] objects are provided.
 #'
-#' The coersion from hclust is a faster reimplementation of [stats::as.dendrogram()].
+#' Coersion methods for
+#' [hclust], [hdbscan] and [reachability] objects to [dendrogram] are provided.
+#'
+#' The coercion from `hclust` is a faster C++ reimplementation of the coercion in
+#' package `stats`. The original implementation can be called
+#' using [stats::as.dendrogram()].
 #'
 #' The coersion from [hdbscan] builds the non-simplified HDBSCAN hierarchy as a
 #' dendrogram object.
@@ -34,29 +41,36 @@
 NULL
 
 #' @rdname dendrogram
-as.dendrogram.hclust <- function(object, ...){
+#' @export
+as.dendrogram <- function (object, ...) {
+  UseMethod("as.dendrogram", object)
+}
+
+#' @rdname dendrogram
+#' @export
+as.dendrogram.default <- function (object, ...)
+  stats::as.dendrogram(object, ...)
+
+## this is a replacement for stats::as.dendrogram for hclust
+#' @rdname dendrogram
+#' @export
+as.dendrogram.hclust <- function(object, ...) {
   return(buildDendrogram(object))
 }
 
 #' @rdname dendrogram
-as.dendrogram.hdbscan <- function(object, ...){
+#' @export
+as.dendrogram.hdbscan <- function(object, ...) {
   return(buildDendrogram(object$hc))
 }
 
 #' @rdname dendrogram
-#' @name reachability-coersion
-#' @aliases as.reachability
-NULL
-
-as.reachability <-
-  function(object, ...)
-    UseMethod("as.reachability")
-
-#' @rdname dendrogram
+#' @export
 as.dendrogram.reachability <- function(object, ...) {
-  if(length(which(object$reachdist == Inf)) > 1) stop("Multiple Infinite reachability distances found. Reachability plots can only be converted if they contain
-                                                      enough information to fully represent the dendrogram structure. If using OPTICS, a larger eps value
-                                                      (such as Inf) may be needed in the parameterization.")
+  if (length(which(object$reachdist == Inf)) > 1)
+    stop(
+      "Multiple Infinite reachability distances found. Reachability plots can only be converted if they contain enough information to fully represent the dendrogram structure. If using OPTICS, a larger eps value (such as Inf) may be needed in the parameterization."
+    )
   #dup_x <- object
   c_order <- order(object$reachdist) - 1
   # dup_x$order <- dup_x$order - 1
@@ -92,7 +106,7 @@ as.dendrogram.reachability <- function(object, ...) {
         depth <- depth + 1L
         if (verbose)
           cat(sprintf(" depth(+)=%4d, k=%d\n", depth,
-                      k))
+            k))
         kk[depth] <- k
         if (storage.mode(jj) != storage.mode(kk))
           storage.mode(jj) <- storage.mode(kk)
@@ -110,12 +124,12 @@ as.dendrogram.reachability <- function(object, ...) {
         depth <- depth - 1L
         if (verbose)
           cat(sprintf(" depth(-)=%4d, k=%d\n", depth,
-                      k))
+            k))
         midS <- sum(vapply(r, .midDend, 0))
         if (!quiet && type == "hclust" && k != 2)
           warning("midcache() of non-binary dendrograms only partly implemented")
         attr(r, "midpoint") <- (.memberDend(r[[1L]]) +
-                                  midS)/2
+            midS) / 2
         d <- r
       }
       if (!depth)
@@ -128,8 +142,12 @@ as.dendrogram.reachability <- function(object, ...) {
   setmid(x, type = type)
 }
 
-.midDend <- function (x)
-  if (is.null(mp <- attr(x, "midpoint"))) 0 else mp
+.midDend <- function (x) {
+  if (is.null(mp <- attr(x, "midpoint")))
+    0
+  else
+    mp
+}
 
 .memberDend <- function (x)
 {