+++ src/R_mrd.cpp
@@ -9,20 +9,22 @@ using namespace Rcpp;
 //
 // Input:
 // * dm: distances as a dist object (vector) of size (n*(n-1))/2 where n
-//       is the number of points
+//       is the number of points.
+//       Note: we divide by 2 early to stay within the number range of int.
 // * cd: the core distances as a vector of length n
 //
 // Returns:
 // a vector (dist object) in the same order as dm
 // [[Rcpp::export]]
 NumericVector mrd(NumericVector dm, NumericVector cd) {
-  int n = cd.length();
-  if (dm.length() != (n*(n-1))/2)
-    stop("number of mutual reachability distance values and size of the distances do not agree.");
+  R_xlen_t n = cd.length();
+  if (dm.length() != (n * (n-1) / 2))
+    stop("number of mutual reachability distance values and size of the distance matrix do not agree.");
 
   NumericVector res = NumericVector(dm.length());
-  for (int i = 0, idx = 0; i < n; ++i) {
-    for (int j = i+1; j < n; ++j, ++idx) {
+  for (R_xlen_t i = 0, idx = 0; i < n; ++i) {
+//    Rprintf("i = %ill of %ill, idx = %ill\n", i, n, idx);
+    for (R_xlen_t j = i+1; j < n; ++j, ++idx) {
       res[idx] = std::max(dm[idx], std::max(cd[i], cd[j]));
     }
   }